add 'create_actor' method and 'to_csv' distr func for dask
